requirements.txt
These requirements were generated using pipreqs

Flask==1.1.4
geopandas==0.10.2
matplotlib==3.3.2
numpy==1.19.2
Pillow==9.2.0
rasterio==1.1.2
Shapely==1.7.1
Werkzeug==1.0.1
